baselayer/requirements.txt
psycopg2==2.7.1
pyyaml
tornado
pyzmq
pyjwt
simplejson
selenium
pytest
social-auth-core
social-auth-app-tornado
-e git://github.com/python-social-auth/social-storage-peewee.git#egg=social-auth-storage-peewee
selenium-requests
